requirements.txt
anyio==4.2.0
astroid==3.0.3
black==24.2.0
certifi==2024.2.2
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.4.1
dill==0.3.8
distlib==0.3.8
filelock==3.13.1
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
identify==2.5.34
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mando==0.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.1
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.0.3
PyMySQL==1.1.0
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.26
tomlkit==0.12.3
typing_extensions==4.9.0
virtualenv==20.25.0
